config/settings.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Twitter API credentials
API_KEY = os.getenv("API_KEY")
API_SECRET_KEY = os.getenv("API_SECRET_KEY")
ACCESS_TOKEN = os.getenv("ACCESS_TOKEN")
ACCESS_TOKEN_SECRET = os.getenv("ACCESS_TOKEN_SECRET")
BEARER_TOKEN = os.getenv("BEARER_TOKEN")

# Bot settings
FOLLOW_THRESHOLD_PERCENT = 10  # Percent of followers needed to follow a new account
REPORT_OUTPUT_PATH = "logs/daily_report.html"
