pyproject.toml
[project]
name = "xsense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tweepy>=4.14.0",
    "pandas>=2.2.3",
    "jinja2>=3.1.5",
    "numpy>=2.2.1",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "loguru>=0.7.3",
    "pytest-bdd>=8.1.0",
    "requests-oauthlib>=1.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
